refactor: consolidate duplicated helpers across ttsc and unplugin
Architecture-level deduplication of helpers that were copy-pasted across
files, with no behavior change:

- ttsc TS: extract the go.mod directory walk (`packageRootDir`,
  `findNearestGoMod`) and `isOutsideRelativePath` into
  `compiler/internal/paths.ts`; the native-binary spawn helpers
  (`spawnNative`, `ensureExecutable`, `outputText`) into
  `compiler/internal/spawnNative.ts`; and `resolveCacheDir` into
  `launcher/internal/resolveCacheDir.ts`. Each was byte-identical across
  its call sites.
- lint Go: `parseExternalRuleEntry` now delegates to the identical
  `parseRuleEntry`.
- unplugin: the shared source-file regex is exported once from
  `core/index.ts` and imported by `bun.ts`.

Duplicates that diverge in behavior were deliberately left separate:
the directive vs external lint-rule-name normalizers (different
trimming and prefix sets) and the trailing-comma vs semi comment-trivia
scanners (different loop-control contracts).

Verified: `node scripts/test-go-lint.cjs` and `pnpm test` both green.
